kalshi/dora_bot/requirements.txt
# Kalshi API
requests>=2.31.0
cryptography>=41.0.0

# AWS
boto3>=1.28.0
botocore>=1.31.0

# Environment
python-dotenv>=1.0.0

# Async
asyncio

# Logging and utilities
